\name{readAsciigrid}
\alias{readAsciigrid}
\alias{writeAsciigrid}
\title{ read/write to/from ESRI asciigrid format }
\description{ read/write to/from ESRI asciigrid format }
\usage{ 
readAsciigrid(fname, as.image = FALSE, plot.image = FALSE)
writeAsciigrid(x, fname, attr, na.value = -9999, ...)
}

\arguments{
\item{fname}{ fname }
\item{as.image}{logical; if FALSE, a list is returned, ready to be
shown with the \code{image} command; if FALSE an object of
class \link{SpatialDataFrameGrid-class} is returned }
\item{plot.image}{logical; if TRUE, an image of the map is plotted}
\item{x}{ object of class \code{SpatialDataFrameGrid}}
\item{attr}{attribute column; if missing, the default (first non-coordinate)
column is taken; a name or a column number may be given}
\item{na.value}{numeric; value given to missing valued cells in the
resulting map}
\item{...}{ arguments passed to \link{write.table}, which is used
to write the numeric data }
}
\value{
	\code{readAsciigrid} returns the grid map read; either as
	an object of class \link{SpatialDataFrameGrid-class} or as list
	with components \code{x}, \code{y} and {z}.
}
\seealso{ \link{as.image.SDFG}, \link{image} }
\author{Edzer J.\ Pebesma, e.pebesma@geog.uu.nl}
\keyword{programming}
