\name{coordnames-methods}
\docType{methods}
\alias{coordnames}
\alias{coordnames-methods}
\alias{coordnames,SpatialPoints-method}
\alias{coordnames,SpatialLines-method}
\alias{coordnames,Slines-method}
\alias{coordnames,Sline-method}
\alias{coordnames,SpatialRings-method}
\alias{coordnames,Srings-method}
\alias{coordnames,Sring-method}
\alias{coordnames<-}
\alias{coordnames<-,SpatialPoints,character-method}
\alias{coordnames<-,SpatialLines,character-method}
\alias{coordnames<-,Slines,character-method}
\alias{coordnames<-,Sline,character-method}
\alias{coordnames<-,SpatialRings,character-method}
\alias{coordnames<-,Srings,character-method}
\alias{coordnames<-,Sring,character-method}
\title{ retrieve or assign coordinate names for classes in sp}
\description{
retrieve or assign coordinate names for classes in \pkg{sp}
}
\section{Methods for coordnames}{
\describe{
\item{x = "SpatialPoints"}{ retrieves coordinate names }
\item{x = "SpatialLines"}{ retrieves coordinate names }
\item{x = "Slines"}{ retrieves coordinate names }
\item{x = "Sline"}{ retrieves coordinate names }
\item{x = "SpatialRings"}{ retrieves coordinate names }
\item{x = "Srings"}{ retrieves coordinate names }
\item{x = "Sring"}{ retrieves coordinate names }
}}
\section{Methods for "coordnames<-"}{
\describe{
\item{x = "SpatialPoints", value = "character"}{ assigns coordinate names }
\item{x = "SpatialLines", value = "character"}{ assigns coordinate names }
\item{x = "Slines", value = "character"}{ assigns coordinate names }
\item{x = "Sline", value = "character"}{ assigns coordinate names }
\item{x = "SpatialRings", value = "character"}{ assigns coordinate names }
\item{x = "Srings", value = "character"}{ assigns coordinate names }
\item{x = "Sring", value = "character"}{ assigns coordinate names }
}}
\keyword{methods}
