\name{meuse.grid_ll}
\alias{meuse.grid_ll}
\docType{data}
\title{Prediction Grid for Meuse Data Set, geographical coordinates}
\description{
  The object contains the meuse.grid data as a SpatialPointsDataFrame 
after transformation to WGS84 and geographical coordinates.
}
\usage{data(meuse.grid_ll)}
\format{
  The format is:
Formal class 'SpatialPointsDataFrame' [package "sp"]. 
}

\source{
  See the \link{meuse} documentation
}
\examples{
data(meuse.grid_ll)
}
\keyword{datasets}
