\name{nc.Map}
\alias{nc.Map}
\alias{ncMap}
\docType{data}
\title{North Carolina Map object}
\description{
  North Carolina Map object
}
\usage{data(ncMap)}
\format{
  The format is class \code{"Map"} from package "maptools".
}

\source{
  Data file in shapefile format included in "maptools" package.
}

\keyword{datasets}
