\name{polygons-methods}
\docType{methods}
\alias{polygons-methods}
\alias{polygons,Spatial-method}
\title{ Retrieve polygons from SpatialPolygonsDataFrame object }
\description{
Retrieve polygons from SpatialPolygonsDataFrame object
}
\section{Methods}{
\describe{
\item{obj = "Spatial"}{ returns SpatialPolygons part if object is of
	class "SpatialPolygonsDataFrame" or "SpatialPolygons"}
%\item{}{ ~~describe this method here }
}}
\keyword{methods}
\keyword{ }
