\name{rings-methods}
\docType{methods}
\alias{rings-methods}
\alias{rings,Spatial-method}
\title{ Retrieve rings from SpatialRingsDataFrame object }
\description{
Retrieve rings from SpatialRingsDataFrame object
}
\section{Methods}{
\describe{
\item{obj = "Spatial"}{ returns SpatialRings part if object is of
	class "SpatialRingsDataFrame" or "SpatialRings"}
%\item{}{ ~~describe this method here }
}}
\keyword{methods}
\keyword{ }
