\name{transform-methods}
\docType{methods}
\alias{transform-methods}
\alias{transform,Spatial-method}
\alias{transform.Spatial}
\title{ place holder for transform methods in library spproj }
\description{
provide useful error message if library spproj is not loaded
}
\section{Methods}{
\describe{
\item{obj = "Spatial"}{ tries to load library spproj and
call the appropriate transform method present there; if
this fails, an error message results. }
}}
\keyword{methods}
\keyword{ }
