\name{coordnames-methods}
\docType{methods}
\alias{coordnames}
\alias{coordnames-methods}
\alias{coordnames,SpatialPoints-method}
\alias{coordnames,SpatialMultiPoints-method}
\alias{coordnames,SpatialLines-method}
\alias{coordnames,Lines-method}
\alias{coordnames,Line-method}
\alias{coordnames,SpatialPolygons-method}
\alias{coordnames,Polygons-method}
\alias{coordnames,Polygon-method}
\alias{coordnames<-}
\alias{coordnames<-,SpatialPoints,character-method}
\alias{coordnames<-,SpatialMultiPoints,character-method}
\alias{coordnames<-,SpatialLines,character-method}
\alias{coordnames<-,Lines,character-method}
\alias{coordnames<-,Line,character-method}
\alias{coordnames<-,SpatialPolygons,character-method}
\alias{coordnames<-,Polygons,character-method}
\alias{coordnames<-,Polygon,character-method}
\alias{coordnames<-,GridTopology,character-method}
\alias{coordnames<-,SpatialGrid,character-method}
\alias{coordnames<-,SpatialPixels,character-method}
\alias{coordnames<-,GridTopology,character-method}
\alias{coordnames<-,SpatialGrid,character-method}
\alias{coordnames<-,SpatialPixels,character-method}
\title{ retrieve or assign coordinate names for classes in sp}
\description{
retrieve or assign coordinate names for classes in \pkg{sp}
}
\section{Methods for coordnames}{
\describe{
\item{x = "SpatialPoints"}{ retrieves coordinate names }
\item{x = "SpatialLines"}{ retrieves coordinate names }
\item{x = "Lines"}{ retrieves coordinate names }
\item{x = "Line"}{ retrieves coordinate names }
\item{x = "SpatialPolygons"}{ retrieves coordinate names }
\item{x = "Polygons"}{ retrieves coordinate names }
\item{x = "Polygon"}{ retrieves coordinate names }
}}
\section{Methods for "coordnames<-"}{
\describe{
\item{x = "SpatialPoints", value = "character"}{ assigns coordinate names }
\item{x = "SpatialLines", value = "character"}{ assigns coordinate names }
\item{x = "Lines", value = "character"}{ assigns coordinate names }
\item{x = "Line", value = "character"}{ assigns coordinate names }
\item{x = "SpatialPolygons", value = "character"}{ assigns coordinate names }
\item{x = "Polygons", value = "character"}{ assigns coordinate names }
\item{x = "Polygon", value = "character"}{ assigns coordinate names }
\item{x = "GridTopology", value = "character"}{ assigns coordinate names }
\item{x = "SpatialGrid", value = "character"}{ assigns coordinate names }
\item{x = "SpatialPixels", value = "character"}{ assigns coordinate names }
}}
\keyword{methods}
