\name{loadMeuse}
\alias{loadMeuse}
\title{ deprecated function to load the Meuse data set }
\description{ deprecated function to load the Meuse data set }
\usage{
loadMeuse()
}
\value{ none; it prints a warning to run demo(meuse) }
\seealso{\link{meuse}, \link{meuse.grid}}
\examples{
demo(meuse)
}
\keyword{dplot}
