\name{Srings-class}
\docType{class}
\alias{Srings-class}

\title{Class "Srings" ~~~ }
\description{ Collection of objects of class \code{"Sring"} }
\section{Objects from the Class}{
Objects can be created by calls to the function \code{Srings}
}
\section{Slots}{
  \describe{
    \item{\code{Srings}:}{Object of class \code{"list"}; list with objects
	of class \link{Sring-class} }
    \item{\code{plotOrder}:}{Object of class \code{"integer"}; order in which
	the Sring objects should be plotted; RSB }
    \item{\code{labpt}:}{Object of class \code{"numeric"}; RSB  }
    \item{\code{ID}:}{Object of class \code{"character"}; RSB  }
    \item{\code{bbox}:}{Object of class \code{"matrix"}; see \link{Spatial-class} }
    \item{\code{proj4string}:}{Object of class \code{"CRS"}; see \link{CRS-class} }
  }
}
\section{Extends}{
Class \code{"Spatial"}, directly.
}
\section{Methods}{
No methods defined with class "Srings" in the signature.
}
\references{ }
\author{ Roger Bivand }
\note{ }

\seealso{
}
\examples{
##---- Should be DIRECTLY executable !! ----
}
\keyword{classes}
