\name{transform-methods}
\docType{methods}
\alias{transform-methods}
\alias{transform,Spatial-method}
\title{ place holder for transform methods in library spproj }
\description{
provide useful error message if library spproj is not loaded
}
\section{Methods}{
\describe{
\item{obj = "Spatial"}{ return error messag; useful transform
methods are provided by library spproj, this method is only
called when spproj is not loaded}
}}
\keyword{methods}
\keyword{ }
